feat: Dto 모델 수정,
