+++ lib/data/dto/weather_dto.dart
@@ -1,0 +1,54 @@
+import 'hourly_dto.dart';
+
+class WeatherDto {
+  WeatherDto({
+    this.latitude,
+    this.longitude,
+    this.generationtimeMs,
+    this.utcOffsetSeconds,
+    this.timezone,
+    this.timezoneAbbreviation,
+    this.elevation,
+    this.hourlyUnits,
+    this.hourly,});
+
+  WeatherDto.fromJson(dynamic json) {
+    latitude = json['latitude'];
+    longitude = json['longitude'];
+    generationtimeMs = json['generationtime_ms'];
+    utcOffsetSeconds = json['utc_offset_seconds'];
+    timezone = json['timezone'];
+    timezoneAbbreviation = json['timezone_abbreviation'];
+    elevation = json['elevation'];
+    hourlyUnits = json['hourly_units'] != null ? HourlyUnits.fromJson(json['hourly_units']) : null;
+    hourly = json['hourly'] != null ? HourlyDto.fromJson(json['hourly']) : null;
+  }
+  num? latitude;
+  num? longitude;
+  num? generationtimeMs;
+  num? utcOffsetSeconds;
+  String? timezone;
+  String? timezoneAbbreviation;
+  num? elevation;
+  HourlyUnits? hourlyUnits;
+  HourlyDto? hourly;
+
+  Map<String, dynamic> toJson() {
+    final map = <String, dynamic>{};
+    map['latitude'] = latitude;
+    map['longitude'] = longitude;
+    map['generationtime_ms'] = generationtimeMs;
+    map['utc_offset_seconds'] = utcOffsetSeconds;
+    map['timezone'] = timezone;
+    map['timezone_abbreviation'] = timezoneAbbreviation;
+    map['elevation'] = elevation;
+    if (hourlyUnits != null) {
+      map['hourly_units'] = hourlyUnits?.toJson();
+    }
+    if (hourly != null) {
+      map['hourly'] = hourly?.toJson();
+    }
+    return map;
+  }
+
+}
